+++ cfgov/v1/models/browse_page.py
@@ -60,9 +60,9 @@ class BrowsePage(CFGOVPage):
 
     template = 'browse-basic/index.html'
 
-    def add_page_js(self, js):
-        super(BrowsePage, self).add_page_js(js)
-        js['template'] += ['secondary-navigation.js']
+    def get_page_js(self):
+        files = super(BrowsePage, self).get_page_js()
+        return files + ['secondary-navigation.js']
 
     def full_width_serif(self):
         return true